fix(amazonq): extend GovCloud block to AgenticChatController (#2777)
* fix(amazonq): extend GovCloud block to AgenticChatController

Follow-up to #2776. The shipped runtime (agent-standalone via
qAgenticChatServer manifest) uses AgenticChatController, not the legacy
ChatController that #2776 patched. Apply the same
getGovCloudUnsupportedResponse guard at the top of
AgenticChatController.onChatPrompt and .onInlineChatPrompt.

* revert(amazonq): remove GovCloud block from legacy ChatController

Scope the GovCloud region block to AgenticChatController only, which is
what the shipped agent-standalone runtime (used by amazon-q-vscode and
JetBrains via the qAgenticChatServer manifest) actually runs. Other
downstream consumers of the legacy ChatController (DEXP, iam-webworker,
etc.) can opt into the same guard on their own timelines.

Leaves getGovCloudUnsupportedResponse and GOV_REGIONS in chat/utils.ts,
since AgenticChatController still consumes them.
